Refactoring structure folder
