+++ src/domain/categories/entities/category.entity.ts
@@ -9,9 +9,21 @@ export class Category {
   updatedAt: Date;
   products: Product[];
 
-  constructor(name, slug, description) {
+  constructor(
+    name: string,
+    slug: string,
+    description: string,
+    id?: string,
+    createdAt?: Date,
+    updatedAt?: Date,
+    products?: Product[],
+  ) {
     this.name = name;
     this.slug = slug;
     this.description = description;
+    this.id = id;
+    this.createdAt = createdAt;
+    this.updatedAt = updatedAt;
+    this.products = products;
   }
 }
