+++ src/domain/customers/customers.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { CustomersService } from './customers.service';
 import { CustomersController } from './adapters/inbound/controller/customers.controller';
 import { ICustomerRepository } from './ports/ICustomerRepository';
-import { CustomerModelRepository } from './adapters/outbound/repositories/customer.repository';
+import { CustomerModelRepository } from '../../frameworks/database/postgres/customers/repositories/customer.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CustomerModel } from './adapters/outbound/models/customer.model';
+import { CustomerModel } from '../../frameworks/database/postgres/customers/models/customer.model';
 import { ICustomersService } from './ports/ICustomersService';
 
 @Module({
