+++ src/domain/customers/entities/customer.entity.ts
@@ -6,9 +6,19 @@ export class Customer {
   createdAt: Date;
   updatedAt: Date;
 
-  constructor(name, email, cpf) {
+  constructor(
+    name: string,
+    email: string,
+    cpf: string,
+    id?: string,
+    createdAt?: Date,
+    updatedAt?: Date,
+  ) {
     this.name = name;
     this.email = email;
     this.cpf = cpf;
+    this.id = id;
+    this.createdAt = createdAt;
+    this.updatedAt = updatedAt;
   }
 }
