+++ src/domain/products/products.module.ts
@@ -2,13 +2,13 @@ import { Module } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { ProductsController } from './adapters/inbound/controllers/products.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ProductModel } from './adapters/outbound/models/product.model';
-import { ProductImageModel } from './adapters/outbound/models/productImage.model';
+import { ProductModel } from '../../frameworks/database/postgres/products/models/product.model';
+import { ProductImageModel } from '../../frameworks/database/postgres/products/models/productImage.model';
 import { ICategoryRepository } from '../categories/ports/ICategoryRepository';
-import { CategoryModelRepository } from '../../framworks/data-services/postgres/categories/repositories/category.repository';
+import { CategoryModelRepository } from '../../frameworks/database/postgres/categories/repositories/category.repository';
 import { IProductRepository } from './ports/IProductRepository';
-import { ProductModelRepository } from './adapters/outbound/repositories/product.repository';
-import { CategoryModel } from '../../framworks/data-services/postgres/categories/models/category.model';
+import { ProductModelRepository } from '../../frameworks/database/postgres/products/repositories/product.repository';
+import { CategoryModel } from '../../frameworks/database/postgres/categories/models/category.model';
 
 @Module({
   imports: [
