+++ src/frameworks/database/postgres/products/models/productImage.model.ts
